refactor(check): extract type-resolution.ts from checker.ts
Extract type resolution functions into a dedicated module to improve
code organization and prepare for the layered architecture refactoring.

Functions extracted:
- Type name resolution: getTypeNameFromToken
- List type resolution: resolveListType and helpers
- User-defined type resolution: resolveUserDefinedType
- Hinted primitive resolution: resolveHintedPrimitive and helpers
- Main type resolution: resolveTypeFromAnnotation
- Refinement constraints: checkRefinementConstraints
- Emission helpers: emitIntConstInst, emitIntBoundsError,
  emitConstraintViolationError, parseIntegerLiteral

This is Task 3 of the checker.ts refactoring plan.
